xe: jit: gemm: split instatiation of BLASKernelGenerator templates
The BLASKernelGenerator template is tens of thousands of lines in size. Because
of this, compilers can take anywhere from 5-30 minutes depending on the machine
and the compiler to compile all templates in a single compilation unit. As this
can be as significant bottleneck on compilation time, this splits each
instantiation into its own library to increase compilation parallelism.